+++ src/main/kotlin/com/github/blarc/gitlab/template/lint/plugin/GitlabLintRunner.kt
@@ -8,13 +8,13 @@ import com.intellij.ui.EditorNotifications
 
 
 fun runLinting(file: PsiFile) {
-    runBackgroundableTask(GitlabLintBundle.message("inspection.title"), file.project) { indicator ->
-        indicator.isIndeterminate = true
-        val project = file.project
-        if (GitlabLintUtils.matchesGitlabLintGlob(file.virtualFile.path)) {
+    if (GitlabLintUtils.matchesGitlabLintGlob(file.virtualFile.path)) {
+        runBackgroundableTask(GitlabLintBundle.message("inspection.title"), file.project) { indicator ->
+            indicator.isIndeterminate = true
+            val project = file.project
             val pipeline = project.service<Pipeline>()
             pipeline.accept(file)
+            EditorNotifications.getInstance(project).updateAllNotifications()
         }
-        EditorNotifications.getInstance(project).updateAllNotifications()
     }
 }