+++ src/main/kotlin/com/github/blarc/gitlab/template/lint/plugin/listeners/FileListener.kt
@@ -1,48 +1,43 @@
 package com.github.blarc.gitlab.template.lint.plugin.listeners
 
 import com.github.blarc.gitlab.template.lint.plugin.GitlabLintUtils
+import com.github.blarc.gitlab.template.lint.plugin.GitlabLintUtils.updateStatusWidget
 import com.github.blarc.gitlab.template.lint.plugin.runLinting
 import com.github.blarc.gitlab.template.lint.plugin.settings.AppSettings
-import com.github.blarc.gitlab.template.lint.plugin.widget.LintStatusWidgetFactory
-import com.intellij.openapi.application.ApplicationManager
+import com.github.blarc.gitlab.template.lint.plugin.widget.LintStatusEnum
+import com.github.blarc.gitlab.template.lint.plugin.widget.LintStatusWidget
 import com.intellij.openapi.components.Service
 import com.intellij.openapi.fileEditor.FileEditorManagerEvent
 import com.intellij.openapi.fileEditor.FileEditorManagerListener
 import com.intellij.openapi.startup.StartupManager
-import com.intellij.openapi.wm.impl.status.widget.StatusBarWidgetSettings
-import com.intellij.openapi.wm.impl.status.widget.StatusBarWidgetsManager
+import com.intellij.openapi.wm.WindowManager
 import com.intellij.psi.PsiManager
 import com.intellij.ui.EditorNotifications
 
 @Service
 class FileListener : FileEditorManagerListener {
 
-    private val lintStatusWidgetFactory = ApplicationManager.getApplication().getService(LintStatusWidgetFactory::class.java)
-    private val statusBarWidgetSettings = ApplicationManager.getApplication().getService(StatusBarWidgetSettings::class.java)
     override fun selectionChanged(event: FileEditorManagerEvent) {
 
         val file = event.newFile ?: return
         val project = event.manager.project
 
-        // Check if file matches regex
-        val matches = GitlabLintUtils.matchesGitlabLintGlob(file.path)
-
         // Hide/show error notification
         EditorNotifications.getInstance(project).updateAllNotifications()
 
-        // Hide/show widget
-        val statusBarWidgetsManager = project.getService(StatusBarWidgetsManager::class.java)
-        if (lintStatusWidgetFactory != null) {
+        // Hide widget if file doesn't match glob
+        if (!GitlabLintUtils.matchesGitlabLintGlob(file.path)) {
+            updateStatusWidget(project, LintStatusEnum.HIDDEN)
+            return
+        }
 
-            statusBarWidgetSettings?.setEnabled(lintStatusWidgetFactory, matches)
-            statusBarWidgetsManager?.updateWidget(LintStatusWidgetFactory::class.java)
-
-            val startupFinished = StartupManager.getInstance(project).postStartupActivityPassed()
-            if (startupFinished && AppSettings.instance.runLintOnFileChange && matches) {
-                // If we can get the psi file, run linting
-                PsiManager.getInstance(project).findFile(file)?.let {
-                    runLinting(it)
-                }
+        // Run linting if startup is finished and linting is enabled
+        val startupFinished = StartupManager.getInstance(project).postStartupActivityPassed()
+        if (startupFinished && AppSettings.instance.runLintOnFileChange) {
+            WindowManager.getInstance().getStatusBar(project).updateWidget(LintStatusWidget.ID)
+            // If we can get the psi file, run linting
+            PsiManager.getInstance(project).findFile(file)?.let {
+                runLinting(it)
             }
         }
     }