+++ src/main/kotlin/com/github/blarc/gitlab/template/lint/plugin/pipeline/Pipeline.kt
@@ -1,20 +1,19 @@
 package com.github.blarc.gitlab.template.lint.plugin.pipeline
 
+import com.github.blarc.gitlab.template.lint.plugin.GitlabLintUtils.updateStatusWidget
 import com.github.blarc.gitlab.template.lint.plugin.gitlab.GitlabLintResponse
 import com.github.blarc.gitlab.template.lint.plugin.pipeline.middleware.*
 import com.github.blarc.gitlab.template.lint.plugin.widget.LintStatusEnum
-import com.github.blarc.gitlab.template.lint.plugin.widget.LintStatusWidget
 import com.intellij.openapi.components.Service
 import com.intellij.openapi.components.service
 import com.intellij.openapi.project.Project
-import com.intellij.openapi.wm.WindowManager
 import com.intellij.psi.PsiFile
 import java.util.*
 
 @Service
 class Pipeline(private val project: Project) {
     var gitlabLintResponse: GitlabLintResponse? = null
-    var lintStatus: LintStatusEnum = LintStatusEnum.WAITING
+    var lintStatus: LintStatusEnum = LintStatusEnum.HIDDEN
 
     private val middlewares: Set<Middleware> = setOf(
         project.service<ResolveContext>(),
@@ -33,11 +32,11 @@ class Pipeline(private val project: Project) {
 
         val queue = PriorityQueue(middlewares)
 
-        updateStatusWidget(LintStatusEnum.RUNNING)
+        updateStatusWidget(project, LintStatusEnum.RUNNING)
         val result = next(queue, Pass(project, file))
 
         gitlabLintResponse = result?.first
-        updateStatusWidget(result?.second ?: LintStatusEnum.INVALID)
+        updateStatusWidget(project,result?.second ?: LintStatusEnum.INVALID)
     }
 
     private fun next(queue: PriorityQueue<Middleware>, pass: Pass) : Pair<GitlabLintResponse?, LintStatusEnum>? {
@@ -46,10 +45,5 @@ class Pipeline(private val project: Project) {
         return middleware(pass) {
             return@middleware next(queue, pass);
         }
-    }
-
-    private fun updateStatusWidget(status: LintStatusEnum) {
-        lintStatus = status
-        WindowManager.getInstance().getStatusBar(project)?.updateWidget(LintStatusWidget.ID)
     }
 }