+++ src/main/kotlin/com/github/blarc/gitlab/template/lint/plugin/widget/LintStatusRepresentation.kt
@@ -5,24 +5,21 @@ import com.github.blarc.gitlab.template.lint.plugin.pipeline.Pipeline
 import com.github.blarc.gitlab.template.lint.plugin.widget.actions.OpenSettingsAction
 import com.github.blarc.gitlab.template.lint.plugin.widget.actions.RefreshAction
 import com.intellij.dvcs.ui.LightActionGroup
-import com.intellij.icons.AllIcons
 import com.intellij.ide.DataManager
 import com.intellij.openapi.actionSystem.ActionGroup
 import com.intellij.openapi.components.service
+import com.intellij.openapi.project.Project
 import com.intellij.openapi.ui.popup.ListPopup
 import com.intellij.openapi.util.Conditions
-import com.intellij.openapi.wm.StatusBar
 import com.intellij.openapi.wm.StatusBarWidget
+import com.intellij.openapi.wm.WindowManager
 import com.intellij.ui.popup.PopupFactoryImpl.ActionGroupPopup
 import javax.swing.Icon
 
-class LintStatusPresentation(private val statusBar: StatusBar) : StatusBarWidget.MultipleTextValuesPresentation {
-    override fun getTooltipText(): String {
-        val status = statusBar.project?.service<Pipeline>()?.lintStatus ?: return LintStatusEnum.WAITING.tooltip
-        return status.tooltip
-    }
+class LintStatusPresentation(val project: Project) : StatusBarWidget.MultipleTextValuesPresentation {
 
     override fun getPopup(): ListPopup {
+        val statusBar = WindowManager.getInstance().getStatusBar(project)
         return ActionGroupPopup(
             GitlabLintBundle.message("lint.status.popup.title"),
             createActions(),
@@ -38,29 +35,27 @@ class LintStatusPresentation(private val statusBar: StatusBar) : StatusBarWidget
         )
     }
 
-    override fun getSelectedValue(): String {
-        val status = statusBar.project?.service<Pipeline>()?.lintStatus ?: return LintStatusEnum.WAITING.text
-        return status.text
-    }
-
-
-    override fun getIcon(): Icon {
-        return when (statusBar.project?.service<Pipeline>()?.lintStatus) {
-            LintStatusEnum.INVALID -> AllIcons.General.ExclMark
-            LintStatusEnum.INVALID_ID -> AllIcons.General.ExclMark
-            LintStatusEnum.VALID -> AllIcons.General.InspectionsOK
-            LintStatusEnum.RUNNING -> AllIcons.General.InlineRefreshHover
-            else -> {
-                AllIcons.General.InspectionsPause
-            }
-        }
-    }
-
     private fun createActions(): ActionGroup {
         val actionGroup = LightActionGroup()
         actionGroup.add(OpenSettingsAction())
         actionGroup.add(RefreshAction())
         return actionGroup
+    }
+
+    override fun getTooltipText(): String? {
+        val status = project.service<Pipeline>().lintStatus
+        return status.tooltip
+    }
+
+
+    override fun getSelectedValue(): String? {
+        val status = project.service<Pipeline>().lintStatus
+        return status.text
+    }
+
+    override fun getIcon(): Icon? {
+        val status = project.service<Pipeline>().lintStatus
+        return status.icon
     }
 
 }